backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "elasticsearch>=9.1.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scalar-fastapi>=1.2.3",
    "uvicorn>=0.35.0",
]
